refactor(honk): lift the evaluator setup into the library
`create_eval_context` and `catch_nock_panic` live in the honk binary, but
they describe the environment *anything* firing honk-minted Nock has to
use: the native hot state with formula-keyed jet dispatch. A second copy
of that setup in another binary would be a silent correctness hazard,
because a divergence surfaces not as a build failure but as a different
noun. Move both into `honk::eval` so there is one definition to change.

`catch_nock_panic` becomes generic over the error type — it built its
error with `.into()` already, so the honk call sites are unchanged.

Two behavioural notes:

  * The eval stack reservation gains a `HONK_EVAL_STACK_WORDS` override,
    in the spirit of the existing `HONK_WORKER_STACK_BYTES`. The 16GB
    default is virtual and free where the kernel overcommits, but a
    process that fires one small trap does not need it, and several at
    once (a test suite) can exceed what a heuristic-overcommit kernel
    will grant. The unit is words, matching `NockStack::new`, so it
    cannot be confused with the byte-denominated worker override.

  * Nothing else changes: same size, same hot state, same dispatch mode.
