+++ crates/honk/src/lib.rs
@@ -11,6 +11,7 @@ pub mod arm_map;
 pub mod artifact;
 pub mod build_cache;
 pub mod errors;
+pub mod eval;
 pub mod nasm_bridge;
 pub mod native;
 pub mod pipeline;
